Refactor: removing extraneous project content
We're after a minimalist example that is not meant to become a full project, so
we hardcode strings inline and use an icon that should exist on the system as
our application icon.  Licensing has been refactored, because all of the remaining
content is either my own or covered by fair use.