New: major update, uses ListFragment and ContentProvider
Major update; hardly any of the original code is left.
This is much more "correct" that the last version, and in fact is acceptable in production modulo bugs.

There may be some behavioral consequences that are off-spec, depending, or the given behavior might be achievable with simpler code.
However, it is mostly meant as a good starting point for developing custom kinds of list views, based on the android ListView.